Modify in Code

+++ app.py
@@ -18,7 +18,7 @@ def home():
 @app.route('/predict', methods=['POST'])
 def predict():
     # Extract data from form
-    int_features = [int(x) for x in request.form.values()]
+    int_features = [float(x) for x in request.form.values()]
     final_features = [np.array(int_features)]
     
     # Make prediction
